GH-788 Move constants into EnerginetRegionConnectorMetadata
