+++ region-connectors/region-connector-dk-energinet/src/main/java/energy/eddie/regionconnector/dk/energinet/EnerginetRegionConnectorMetadata.java
@@ -3,10 +3,15 @@ package energy.eddie.regionconnector.dk.energinet;
 import energy.eddie.api.v0.RegionConnectorMetadata;
 
 import javax.annotation.Nullable;
+import java.time.Period;
+import java.time.ZoneId;
 
 public class EnerginetRegionConnectorMetadata implements RegionConnectorMetadata {
     public static final String REGION_CONNECTOR_ID = "dk-energinet";
-    public static final String BASE_PATH = "/region-connectors/" + REGION_CONNECTOR_ID;
+    public static final ZoneId DK_ZONE_ID = ZoneId.of("Europe/Copenhagen");
+    public static final Period PERIOD_EARLIEST_START = Period.ofYears(-2);
+    public static final Period PERIOD_LATEST_END = Period.ofYears(2);
+    public static final long MAXIMUM_MONTHS_IN_THE_PAST = Math.abs(PERIOD_EARLIEST_START.toTotalMonths());
     @Nullable
     private static EnerginetRegionConnectorMetadata instance = null;
