+++ region-connectors/region-connector-dk-energinet/src/main/java/energy/eddie/regionconnector/dk/energinet/permission/request/states/EnerginetCustomerCreatedState.java
@@ -6,7 +6,6 @@ import energy.eddie.api.agnostic.process.model.validation.AttributeError;
 import energy.eddie.api.agnostic.process.model.validation.ValidationException;
 import energy.eddie.api.agnostic.process.model.validation.Validator;
 import energy.eddie.api.v0.PermissionProcessStatus;
-import energy.eddie.regionconnector.dk.energinet.EnerginetRegionConnector;
 import energy.eddie.regionconnector.dk.energinet.permission.request.StateBuilderFactory;
 import energy.eddie.regionconnector.dk.energinet.permission.request.api.DkEnerginetCustomerPermissionRequest;
 import energy.eddie.regionconnector.dk.energinet.permission.request.validation.NotOlderThanValidator;
@@ -17,11 +16,13 @@ import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.Set;
 
+import static energy.eddie.regionconnector.dk.energinet.EnerginetRegionConnectorMetadata.MAXIMUM_MONTHS_IN_THE_PAST;
+
 public class EnerginetCustomerCreatedState
         extends ContextualizedPermissionRequestState<DkEnerginetCustomerPermissionRequest>
         implements CreatedPermissionRequestState {
     private static final Set<Validator<DkEnerginetCustomerPermissionRequest>> VALIDATORS = Set.of(
-            new NotOlderThanValidator(ChronoUnit.MONTHS, EnerginetRegionConnector.MAXIMUM_MONTHS_IN_THE_PAST),
+            new NotOlderThanValidator(ChronoUnit.MONTHS, MAXIMUM_MONTHS_IN_THE_PAST),
             new CompletelyInThePastOrInTheFutureValidator<>(),
             new StartIsBeforeOrEqualEndValidator<>()
     );
