+++ region-connectors/region-connector-dk-energinet/src/main/java/energy/eddie/regionconnector/dk/energinet/services/IdentifiableApiResponseService.java
@@ -13,7 +13,7 @@ import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Optional;
 
-import static energy.eddie.regionconnector.dk.energinet.EnerginetRegionConnector.DK_ZONE_ID;
+import static energy.eddie.regionconnector.dk.energinet.EnerginetRegionConnectorMetadata.DK_ZONE_ID;
 
 @Service
 public class IdentifiableApiResponseService {
