+++ region-connectors/region-connector-dk-energinet/src/main/java/energy/eddie/regionconnector/dk/energinet/services/PollingService.java
@@ -25,7 +25,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-import static energy.eddie.regionconnector.dk.energinet.EnerginetRegionConnector.DK_ZONE_ID;
+import static energy.eddie.regionconnector.dk.energinet.EnerginetRegionConnectorMetadata.DK_ZONE_ID;
 
 @Service
 public class PollingService implements AutoCloseable {
