+++ region-connectors/region-connector-dk-energinet/src/main/java/energy/eddie/regionconnector/dk/energinet/utils/ConsumptionRecordMapper.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import static energy.eddie.regionconnector.dk.energinet.EnerginetRegionConnector.DK_ZONE_ID;
+import static energy.eddie.regionconnector.dk.energinet.EnerginetRegionConnectorMetadata.DK_ZONE_ID;
 
 public class ConsumptionRecordMapper {
     private ConsumptionRecordMapper() {
